+++ SistemaFauracao/src/main/java/tinario9945/gmail/com/SistemaFauracao/Repository/FaturaRepository.java
@@ -1,10 +1,13 @@
 package tinario9945.gmail.com.SistemaFauracao.Repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import tinario9945.gmail.com.SistemaFauracao.Models.Fatura;
+
 @Repository
-public interface FaturaRepository extends JpaRepository<Fatura,Long>{
-    
+public interface FaturaRepository extends JpaRepository<Fatura,Integer>{
+       
 }
